Revert "Merge pull request #175 from vshalgarg/backend/bug-fix"
This reverts commit 67dae87f307eb9afc4a989576eb7c22f05bcdfb8, reversing
changes made to 4df275219d4680955a178bd88031b7164b8cd509.

+++ src/main/java/com/code/monks/csm/dto/request/BillEntryRequestDto.java
@@ -6,7 +6,6 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -29,8 +28,8 @@ public class BillEntryRequestDto {
     private String lrNumber;
     private String remarks;
 
-    private BigDecimal taxableValue;
-    private BigDecimal billAmount;
+    private double taxableValue;
+    private double billAmount;
 
     @NotEmpty(message = "At least one bill item is required")
     @Valid
