+++ src/main/java/com/code/monks/csm/dto/request/BillUpdateRequest.java
@@ -3,7 +3,6 @@ package com.code.monks.csm.dto.request;
 import com.code.monks.csm.dto.response.BillItemDto;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -18,8 +17,8 @@ public class BillUpdateRequest {
     private String transport;
     private String lrNumber;
     private String remarks;
-    private BigDecimal taxableValue;
-    private BigDecimal billAmount;
+    private Double taxableValue;
+    private Double billAmount;
     private List<BillItemDto> billItems;
     private List<String> existingImageKeys;
 }
